// cpu-app/ClientApp/src/app/core/api/models/documentItemDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */

export interface DocumentItemDto {
  /** @nullable */
  activitymimeattachmentid?: string | null;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  subjectOther?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  overwritetime?: string | null;
}