// cpu-app/ClientApp/src/app/core/api/models/signedContractPostFromPortal.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */
import type { DynamicsDocumentPost } from './dynamicsDocumentPost';
import type { DynamicsSignaturePost } from './dynamicsSignaturePost';

export interface SignedContractPostFromPortal {
  /** @nullable */
  businessBCeID?: string | null;
  /** @nullable */
  userBCeID?: string | null;
  /** @nullable */
  documentCollection?: DynamicsDocumentPost[] | null;
  signature?: DynamicsSignaturePost;
  isModificationAgreement?: boolean;
}